requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Deep Learning & Transformers
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0

# Data Processing
datasets>=2.0.0
huggingface-hub>=0.10.0

# API & LLM
google-generativeai>=0.3.0
requests>=2.28.0
tenacity>=8.0.0

# Visualization & Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Text Processing
nltk>=3.7
regex>=2022.1.18

# Utilities
python-dotenv>=0.19.0
jsonschema>=4.0.0
pyyaml>=6.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipywidgets>=7.6.0
